Add Pandora's box routing
- Cleaned up prev changes a bit
- Added what I like to call Pandora's box routing:
Nodes are able to write network payloads directly to any host, who will
then route the payloads as required. If sending to the host that will
receive the payload, a radio ACK will be sent directly to the sender. If
sending to a host that is not the recipient, it will be forwarded to the
correct host, and a network ACK will be sent by the final delivering
node, The network ack will traverse the network just as any other
message, and will fail if the main link to the sending node is down.
This allows failover nodes to be put in place. Failover nodes should not
have any children, OR should have spare pipes available. Node 014 will
always write to pipe1 for pandora routing, and node 023 will always
write to pipe 2, so any failover nodes should have pipes 1 and 2
available if used for both.
Duplicate payloads are still an issue when network acks fail and are
retried, so can be filtered out by the receiver based on the network
header id number if required.